requirements.txt
absl-py==2.1.0
accelerate==0.26.1
aiohttp==3.9.3
aiosignal==1.3.1
angle-emb==0.3.1
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.42.0
boltons==23.1.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.3
datasets==2.16.1
debugpy==1.8.0
deepspeed==0.13.1
dill==0.3.7
distro==1.9.0
docker-pycreds==0.4.0
docstring-parser==0.15
einops==0.7.0
evaluate==0.4.1
exceptiongroup==1.2.0
# faiss==1.7.4
faiss-gpu==1.7.2
Farama-Notifications==0.0.4
fast-bleu==0.0.90
fastapi==0.109.0
filelock==3.13.1
# flash-attn==2.5.7
fonttools==4.47.2
frozenlist==1.4.1
fschat==0.2.35
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.27.0
google-auth-oauthlib==1.2.0
greenlet==3.0.3
grpcio==1.60.0
gymnasium==0.29.1
h11==0.14.0
hjson==3.1.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
Jinja2==3.1.3
joblib==1.3.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.1.5
langchain-community==0.0.17
langchain-core==0.1.18
langsmith==0.0.85
lightning-utilities==0.10.1
Markdown==3.5.2
markdown-it-py==3.0.0
markdown2==2.4.12
MarkupSafe==2.1.4
marshmallow==3.20.2
matplotlib==3.8.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
networkx==3.2.1
nh3==0.2.15
ninja==1.11.1.1
nltk==3.8.1
numpy==1.26.3
# nvidia-cublas-cu12==12.1.3.1
# nvidia-cuda-cupti-cu12==12.1.105
# nvidia-cuda-nvrtc-cu12==12.1.105
# nvidia-cuda-runtime-cu12==12.1.105
# nvidia-cudnn-cu12==8.9.2.26
# nvidia-cufft-cu12==11.0.2.54
# nvidia-curand-cu12==10.3.2.106
# nvidia-cusolver-cu12==11.4.5.107
# nvidia-cusparse-cu12==12.1.0.106
# nvidia-nccl-cu12==2.19.3
# nvidia-nvjitlink-cu12==12.3.101
# nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openai==1.10.0
optree==0.10.0
packaging==23.2
pandas==2.2.0
peft==0.8.2
pillow==10.2.0
prettytable==3.9.0
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.14
pydantic_core==2.16.1
Pygments==2.17.2
pynvml==11.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.4
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rich==13.7.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
sentence-transformers==2.6.1
sentencepiece==0.1.99
sentry-sdk==1.40.0
setproctitle==1.3.3
shortuuid==1.0.11
shtab==1.7.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
svgwrite==1.4.3
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.15.2
# torch==2.1.2+cu118
# torchaudio==2.1.2+cu118
# torchvision==0.16.2+cu118
torchmetrics==1.3.0.post0
tqdm==4.66.1
transformers==4.38.2
triton==2.1.0
trl==0.8.5
typing-inspect==0.9.0
typing_extensions==4.9.0
tyro==0.7.3
tzdata==2023.4
urllib3==2.2.0
uvicorn==0.27.0.post1
vendi-score==0.0.3
wandb==0.16.6
wavedrom==2.0.3.post3
wcwidth==0.2.13
Werkzeug==3.0.1
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
